+++ app/Http/Livewire/Profile.php
@@ -14,6 +14,7 @@ class Profile extends Component
         'user.email' => 'required|email',
         'user.format_isbns_enabled' => 'boolean',
         'user.date_format' => 'required',
+        'user.secondary_title_preference' => 'required|integer|min:0',
     ];
 
     public function mount(): void