+++ app/Services/SeriesService.php
@@ -31,6 +31,9 @@ class SeriesService
         $series->total = $apiSeries['total'];
         $series->default_price = $apiSeries['default_price'];
         $series->image_url = $apiSeries['image_url'];
+        $series->source_status = $apiSeries['source_status'];
+        $series->source_name = $apiSeries['source_name'];
+        $series->source_name_romaji = $apiSeries['source_name_romaji'];
 
         $publisher = Publisher::whereName($apiSeries['publisher'])->first();
         if (!empty($publisher)) {